Add schematics for the adapter between ioio and the induction hob
A shift register keeps track of the shift register in the left
keyboard/display. The IOIO samples the value when STCP (RCLK)
goes high. The shift register on the right keyboard/display
seems to get the same values but with Q3 (QD) (connected to
INH on the 4051) inverted. Thise makes it possible to emulate
key presses on the right keybaord/display as well.