Fix clock press timing
If the user has pressed the clock we must wait for the hob to leave
the "clock mode" before trying to simulate presses otherwise we might
end up changing the time of the clock instead of the power level. This
almost worked, but did not take into consideration that the hob stays
in "clock mode" while adjusting the time. Solve this by only clearing
the "clock mode" once we have got ok to press a button.